backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-rq==1.16.0
redis==5.0.1
httpx==0.26.0
playwright==1.41.0
beautifulsoup4==4.12.3
slowapi==0.1.9
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
lxml==5.1.0
